First pass at supporting RC6 codes > 32bit (e.g. 36bit MCE/Xbox codes)
* Use 'unsigned long long's in RC6 and decode_results to handle up to 64 bit codes
* Make a temp variable unsigned.
* Add a safety check to RC6 sending when sending more bits than supplied.

+++ IRremoteESP8266.h
@@ -77,7 +77,7 @@ public:
     unsigned int panasonicAddress;
     unsigned int sharpAddress;
   };
-  unsigned long value; // Decoded value
+  unsigned long long value; // Decoded value
   int bits; // Number of bits in decoded value
   volatile unsigned int *rawbuf; // Raw intervals in .5 us ticks
   int rawlen; // Number of records in rawbuf.
@@ -192,7 +192,7 @@ public:
   void sendRaw(unsigned int buf[], int len, int hz);
   void sendGC(unsigned int buf[], int len);
   void sendRC5(unsigned long data, int nbits);
-  void sendRC6(unsigned long data, int nbits);
+  void sendRC6(unsigned long long data, int nbits);
   void sendDISH(unsigned long data, int nbits);
   void sendSharp(unsigned int address, unsigned int command);
   void sendSharpRaw(unsigned long data, int nbits);
